pyproject.toml
[project]
name = "core_wms"
version = "0.1.0"
description = ""
authors = [
    {name = "Paddy",email = "paddy.natte@upfront.nl"}
]
packages = [{ include = "core_wms" }]

requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "strawberry-graphql[subscriptions] (>=0.275.5,<0.276.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "boto3 (>=1.38.41,<2.0.0)",
    "redis (>=5.0.0,<6.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "mypy (>=1.17.0,<2.0.0)",
    "google-cloud-pubsub (>=2.31.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "dramatiq[redis,watch] (>=1.18.0,<2.0.0)",
    "arq (>=0.26.3,<0.27.0)"
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 79
target-version = "py311"
lint.select = ["E", "F", "UP"]
fix = true
[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true